Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -56,12 +56,12 @@ def get_occupation(data, hometown)
 end
 
 def get_average_age_for_season(data, season)
-  total_sum = 0
+  total_sum = 0.0
   counter = 0
   data.each do |correct_season, contestants_info|
     if correct_season == season
     contestants_info.each do |contestant|
-      total_sum = total_sum + contestant["age"].to_i
+      total_sum = total_sum + contestant["age"].to_f
       counter = counter + 1
     end
     end
